test: merge combo-box item renderer and class name generator tests (#10241)
