+++ packages/multi-select-combo-box/test/chips.test.js
@@ -567,58 +567,4 @@ describe('chips', () => {
       expect(overlayPart.clientWidth).to.be.equal(comboBox.clientWidth);
     });
   });
-
-  describe('itemClassNameGenerator', () => {
-    beforeEach(() => {
-      comboBox.autoExpandHorizontally = true;
-    });
-
-    it('should set class name on the selected item chips', async () => {
-      comboBox.itemClassNameGenerator = (item) => item;
-      comboBox.selectedItems = ['apple', 'lemon'];
-      await nextRender();
-
-      const chips = getChips(comboBox);
-      expect(chips[1].className).to.equal('apple');
-      expect(chips[2].className).to.equal('lemon');
-    });
-
-    it('should set class name when generator set after selecting', async () => {
-      comboBox.selectedItems = ['apple', 'lemon'];
-      await nextRender();
-
-      comboBox.itemClassNameGenerator = (item) => item;
-      await nextRender();
-
-      const chips = getChips(comboBox);
-      expect(chips[1].className).to.equal('apple');
-      expect(chips[2].className).to.equal('lemon');
-    });
-
-    it('should remove class name when generator returns empty string', async () => {
-      comboBox.itemClassNameGenerator = (item) => item;
-      comboBox.selectedItems = ['apple', 'lemon'];
-      await nextRender();
-
-      comboBox.itemClassNameGenerator = () => '';
-      await nextRender();
-
-      const chips = getChips(comboBox);
-      expect(chips[1].className).to.equal('');
-      expect(chips[2].className).to.equal('');
-    });
-
-    it('should remove class name when generator is set to null', async () => {
-      comboBox.itemClassNameGenerator = (item) => item;
-      comboBox.selectedItems = ['apple', 'lemon'];
-      await nextRender();
-
-      comboBox.itemClassNameGenerator = null;
-      await nextRender();
-
-      const chips = getChips(comboBox);
-      expect(chips[1].className).to.equal('');
-      expect(chips[2].className).to.equal('');
-    });
-  });
 });
